Clean build with gcc (no warnings)

+++ tests/lbpm_color_simulator.cpp
@@ -389,7 +389,7 @@ int main(int argc, char **argv)
 	FILE *IDFILE = fopen(LocalRankFilename,"rb");
 	if (IDFILE==NULL) ERROR("Error opening file: ID.xxxxx");
 	readID=fread(id,1,N,IDFILE);
-    if (readID != N) printf("lbpm_segmented_pp: Error reading ID (rank=%i) \n",rank);
+	if (readID != size_t(N)) printf("lbpm_segmented_pp: Error reading ID (rank=%i) \n",rank);
 
 	fclose(IDFILE);
